pyproject.toml
[tool.poetry]
name = "university"
version = "0.1.0"
description = ""
authors = ["Skydesperado <Skydesperado@iCloud.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
pyyaml = "^6.0.1"
pytest = "^7.4.3"
pytest-xdist = "^3.3.1"
pytest-django = "^4.7.0"
model-bakery = "^1.17.0"
drf-spectacular = "^0.26.5"
redis = "^5.0.1"
celery = "^5.3.5"
psycopg2-binary = "^2.9.9"
flower = "^2.0.1"
pillow = "^10.1.0"
django-redis = "^5.4.0"
django-rosetta = "^0.9.9"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
djangorestframework-simplejwt = "^5.3.0"
pika = "^1.3.2"
django-minio-storage = "^0.5.7"
django-anymail = "^10.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
flake8 = "^6.1.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
